requirements.txt
flask==1.1.2
python-dotenv==.14.0
pymongo[srv,tls]
pymongo==3.10.1
flask-jwt-extended==3.24.1